[FEATURE] Add review command for Flow work
review <command>
review <changeid>[ <changeid>]*

The review command finds the root of the flow distribution and then runs
any applicable `git gerrit <command>` for each package that needs it.

Available commands: init, update, reset

When one or more <changeid>s are provided, this queries gerrit to find
out which project to apply them to, and then in the project folder
runs `git gerrit apply <changeid>`. Each <changeid> is applied in
succession, so if you select multiple changeids for the same project,
be sure to correctly order them according to dependencies.

Potential future work: Add `review topic:<topic>` to apply all the
changes in a gerrit topic.