[BUGFIX] Make review work with other flow distros
Originally, this assumed that you would be using flow-base-distribution
or neos-base-distribution, so it tried to get the gerrit server info
from the distribution. But that assumption is fragile at best. Now, it
looks for the gerrit server info first in the distribution, then in the
cwd, then it assumes you're working with a git.typo3.org Package, and so
switches to Packages/Framework/TYPO3.Flow and grabs the gerrit server
info from there.

Also, this fixes the case where the distribution directory is a symlink.